release version 0.4.1 2024-05-22
Now using the term slot and submission in place of entry as the
IOCCC uses the term entry to refer to submissions that won an IOCCC.

Changed the user level `entries.json` to `slots.json` as
per the above change.

Start of a mechanism to lock a user slot (via filelock) when operating
on the user slot directory.  As of this release the mechanism is not
being used.  This will allow shell scripts using the `flock(1)`
to not step on a user submitting or updating their IOCCC submission.

TODO: Replace the user directory level `slots.json` (that was
      `entries.json`) with a per user slot directory level `slot.json`
      file.  When writing the user slot directory level `slot.json` file,
      write first to a temporary file in the user slot directory and then
      rename the temporary file to `slot.json`.

TODO: Use the lock mechanism to lock a user slot whenever it
      is being read or written to.

TODO: Add _filename_, _length_, _timestamp_, _SHA256_, and _remark_
      properties to the user slot directory level `slot.json` file.  Print
      non-null information when showing the slot table on the web site.