added password list for initial password generation
Updated `Dockerfile` to no longer use aspell to generate a
dictionary word list.

Removed from `Dockerfile`, commands that were not necessary.  Removed use
of `chmod(1)` in `Dockerfile` as the files under git have their proper
permission.  Lock files are under git and do not need to be created in
the `Dockerfile`.  Directories with files under git do not need to be
created in the `Dockerfile`.

Changed `Dockerfile` VERSION from "0.6 2024-11-22"
to "0.7 2024-11-23".

Tested `bin/ioccc.py` under the command line as well as via Docker.