Format and typo fixes in README.md
The file favicon.ico was added (a copy of the IOCCC website favicon.ico)
so that a 404 is not generated when opening the page.

The file ioccc-submit-index-page.jpg was added to show what the web page
should look like.

The .gitignore has .DS_Store added to it.

Moved indented commands in command blocks to first column as that is an
IOCCC website thing for later use and GitHub should now show the text in
the first column to make it slightly easier to see longer lines.

Clarified and typo checked (up to at least the part where I can check; I
could not check all because docker is deprecated and part of the
documentation only works if docker is running - as I have not put it on
my web server due to that being a production server).

There might have been some other changes in README.md but until the
docker in documentation has been corrected (to refer to how it should be
done now that docker is deprecated) I cannot easily verify that things
are good.