add initial Makefile
We add a `Makefile` to perform some of the python package build actions.

We we know that this is not "Pythonic" but python needs to
learn to "play well others".  :-)

The `template.setup.cfg` file is used to put the value of
the make variable `${VERSION}` into `setup.cfg` when building
that file.

We remove `setup.cfg` from the repo and add it to
`.gitignore` because the new `Makefile`.

Ran `make clobber all` to test.